Making previous and next issue buttons conditional
Conditional on there being a next or previous issue that is!
An array is generated with all the issue numbers in it and if an issue is included in the array then the buttons exist.
Also noted that the markup for two of the issues had the incorrect magazine cover image.