Adding the ebay smart Placement ad code
The reason this wasn't working is to do with how gatsby's Link component loads the page it links to in the background to reduce loading time. However it doesn't work with external links and the ebay script is enough of an external link to prevent that banner ad from preloading. Therefore when we visit the link the banner ad isn't there.