feat: Adding buttons for switching between issues
Adding a panel for containing the Next and Previous Issue buttons.
Adding a common link button component.
Adding a strings file for a single source of constant strings in the project, such as Next Issue and the common slug for issues.
Styling of the above and adding to Issue Template.